PatcheAppleAzaliaAudio - This driver overrides the activateAudioEngine method to add a (modified) Mini9MuteControl class (taken from the patched IOAudioFamily). - This will enable and disable speaker output as required including before and after sleep. - I have removed the power management code from Mini9MuteController as the only real use it has was to initialize speaker output which I have done by manually calling updateMuteContol() in the activateAudioEngine method.
PatchedAppleHDA
- This kext is the same as above although it will only work on unpatched AppleHDA kexts (for the time being). This has not been extensively tested and code needs to be copied over from the above kext. At the moment, this kext will probably cause kernel panics and just not do anything.

And finally, the AppleHDADriver.h and AppleAzaliaAudioDriver.h files are not complete and  have incorrect return type. If you ever plan on using them, verify that they are correct and text the return types. Also, method variables are currently not included in the header files, this may change in the future if they are going to be used (or if I update the header generator program).
Also, the AppleHDa driver was generated form the 10.5.8 kext (I believe) and the Azalia header was generated from the extension included w/ NetbookInstaller.